requirements.txt
python-telegram-bot==21.6
loguru==0.7.2
python-dotenv==1.0.1
python-docx==1.1.2